+++ kevm-pyk/src/tests/conftest.py
@@ -15,8 +15,19 @@ def pytest_addoption(parser: Parser) -> None:
         default=False,
         help='Write expected output files for proof tests',
     )
+    parser.addoption(
+        '--use-booster',
+        action='store_true',
+        default=False,
+        help='Use the kore-rpc-booster binary instead of kore-rpc',
+    )
 
 
 @pytest.fixture
 def update_expected_output(request: FixtureRequest) -> bool:
     return request.config.getoption('--update-expected-output')
+
+
+@pytest.fixture(scope='module')
+def use_booster(request: FixtureRequest) -> bool:
+    return request.config.getoption('--use-booster')
